Moved the loader spy into a separate file
